critical fix for partition table load routine
When loading the partition table in LBA mode the CHS values get
calculated from the LBA sector numbers. The end CHS values were
calculated in the form: start_sector + num_sectors which is wrong,
because it is one too high. The last sector of the partition is
start_sector + num_sectors - 1.