add /NOIPL argument to prevent writing IPL on MBR initialization
IPL area is filled with zero instead. Some BIOSes contain heuristics
which allow them to skip the device upon boot if they detect the
first bytes of the IPL being zero.

FreeDOS 1.3 installation under VMware accidentally works because of
this, because FDISK 1.3 fails to write proper boot code when
initializing the MBR.

The INT 18 method introduced to the boot code via 301ba27 should render
this obsolete.